+++ static_src/components/form/index.js
@@ -1,0 +1,15 @@
+import Form from './form.jsx';
+import FormElement from './form_element.jsx';
+import FormError from './form_error.jsx';
+import FormNumber from './form_number.jsx';
+import FormSelect from './form_select.jsx';
+import FormText from './form_text.jsx';
+
+export {
+  Form,
+  FormElement,
+  FormError,
+  FormNumber,
+  FormSelect,
+  FormText
+};
